fix: Null 예외 원인 제거
