+++ src/main/java/christmas/domain/Benefits.java
@@ -1,6 +1,12 @@
 package christmas.domain;
 
 
+import static christmas.global.Badge.NONE;
+import static christmas.global.Badge.SANTA;
+import static christmas.global.Badge.STAR;
+import static christmas.global.Badge.TREE;
+
+import christmas.global.Badge;
 import christmas.global.Event;
 import java.util.Map;
 
@@ -19,6 +25,31 @@ public class Benefits {
                     .sum();
         }
         return 0;
+    }
+
+    public Badge toEventBadge(int totalBenefit) {
+        if (isStarBadge(totalBenefit)) {
+            return STAR;
+        }
+        if (isTreeBadge(totalBenefit)) {
+            return TREE;
+        }
+        if (isSantaBadge(totalBenefit)) {
+            return SANTA;
+        }
+        return NONE;
+    }
+
+    private boolean isStarBadge(int totalBenefit) {
+        return totalBenefit >= STAR.getAmount() && totalBenefit < TREE.getAmount();
+    }
+
+    private boolean isTreeBadge(int totalBenefit) {
+        return totalBenefit >= TREE.getAmount() && totalBenefit < SANTA.getAmount();
+    }
+
+    private boolean isSantaBadge(int totalBenefit) {
+        return totalBenefit >= SANTA.getAmount();
     }
 
     public Map<Event, Integer> getBenefits() {
