+++ src/main/java/christmas/global/Badge.java
@@ -3,7 +3,8 @@ package christmas.global;
 public enum Badge {
     STAR("별", 5_000),
     TREE("트리", 10_000),
-    SANTA("산타", 20_000);
+    SANTA("산타", 20_000),
+    NONE("없음", 0);
 
     private final String type;
     private final int amount;
